Move display setup & startup tasks out of anaconda.py
Move the display setup functions and some startup
related code out of the main anaconda.py file.
